Etap 0.1

+++ src/Algorytm_Genetyczny.h
@@ -2,4 +2,6 @@
 #include "Macierze.h"
 #include "Osobnik.h"
 
-Osobnik uruchom_GA(const Macierz& macierz, int rozmiar_populacji, int max_pokolen, double wsp_mutacji, double wsp_krzyzowania);
+#include <cstddef>
+
+Osobnik uruchom_GA(const Macierz& macierz, size_t rozmiar_populacji, int max_pokolen, double wsp_mutacji, double wsp_krzyzowania);
